[Feat] 예약 관리 구조체 추가
