+++ BirdCinema/Models/ReservationData.swift
@@ -1,0 +1,17 @@
+//
+//  ReservationData.swift
+//  BirdCinema
+//
+//  Created by 정유진 on 2024/04/24.
+//
+
+import Foundation
+
+struct ReservationData {
+    var movieTitle: String
+    var date: String
+    var time: String
+    var adultCount: Int
+    var youthCount: Int
+    var totalPrice: Int
+}
